pyproject.toml
[tool.poetry]
name = "ntdsoftware"
version = "0.1.0"
description = ""
authors = ["Nahuel Rossi <nahuelrossi.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
requests = "^2.31.0"
django-mysql = "^4.11.0"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
